Rename 'video' package to 'capturer'
