+++ webrtc-kmp/src/iosArm64Main/kotlin/com/shepeliev/webrtckmp/capturer/CameraVideoCapturerController.iosArm64.kt
@@ -1,4 +1,4 @@
-package com.shepeliev.webrtckmp.video
+package com.shepeliev.webrtckmp.capturer
 
 import WebRTC.RTCCameraVideoCapturer
 import WebRTC.RTCVideoCapturerDelegateProtocol
