+++ webrtc-kmp/src/iosMain/kotlin/com/shepeliev/webrtckmp/capturer/VideoCapturerController.kt
@@ -1,4 +1,4 @@
-package com.shepeliev.webrtckmp.video
+package com.shepeliev.webrtckmp.capturer
 
 import com.shepeliev.webrtckmp.MediaTrackSettings
