+++ webrtc-kmp/src/iosX64AndSimulatorArm64Main/kotlin/com/shepeliev/webrtckmp/capturer/CameraVideoCapturerController.iosSimulator.kt
@@ -1,4 +1,4 @@
-package com.shepeliev.webrtckmp.video
+package com.shepeliev.webrtckmp.capturer
 
 import WebRTC.RTCFileVideoCapturer
 import WebRTC.RTCLogEx
